+++ src/main/java/roomescape/global/config/AdminAuthorizationInterceptor.java
@@ -1,4 +1,4 @@
-package roomescape.globar.config;
+package roomescape.global.config;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
