+++ src/main/java/roomescape/global/exception/ReservationExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.globar.exception;
+package roomescape.global.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 @RestControllerAdvice
 public class ReservationExceptionHandler {
 
-    @ExceptionHandler
-    protected ResponseEntity<ErrorResponse> handleDefaultException(RuntimeException e) {
-        return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
-    }
+  @ExceptionHandler
+  protected ResponseEntity<ErrorResponse> handleDefaultException(final RuntimeException e) {
+    return ResponseEntity.badRequest().body(new ErrorResponse(e.getMessage()));
+  }
 }
